+++ nist.py
@@ -6,13 +6,51 @@ and generates a Fortran module.
 Contributed by Vincent MAGNIN, 2021-03-11
 MIT license
 https://github.com/vmagnin/fundamental_constants
-Last modifications: 2021-03-13
+Last modifications: 2021-03-20
 '''
 
 import argparse
 import os
 import datetime
 from urllib.request import urlopen
+
+def get_number_parts(xstring):
+    '''
+    Receives a floating point passed as a string, and returns
+    a tuple with the digits at the left and right of the
+    decimal point, and the value of the exponent.
+    '''
+    # Is there an exponent?
+    exponent_pos = xstring.find("e")
+    if exponent_pos != -1:
+        expo = int(xstring[exponent_pos+1:])
+        # We remove the exponent
+        # (right bound excluded in Python):
+        xstring = xstring[0:exponent_pos]
+    else:
+        expo = 0
+
+    # Number of digits at the left of the decimal point:
+    point_pos = xstring.find(".")
+    if point_pos == -1:
+        left = len(xstring)
+    else:
+        left = point_pos
+    # Must be corrected if a sign is present:
+    if xstring[0] in ('-', '+'):
+        left = left - 1
+
+    # We keep only the decimals and count them:
+    if point_pos != -1:
+        xstring = xstring[point_pos+1:]
+        right = len(xstring)
+    else:
+        right = 0
+
+    return [left, right, expo]
+
+#print(get_number_parts("-7294.29954142"))
+#quit()
 
 # Command line arguments:
 PARSARG = argparse.ArgumentParser(description="""Downloads CODATA recommended 
@@ -67,6 +105,15 @@ f_file.write("\nmodule " + MODULE_NAME + "\n" + TABS +
              "use, intrinsic :: iso_fortran_env, only: wp=>real64\n" +
              TABS + "implicit none\n\n")
 
+# Creating a test file
+test_file = open("test/test.f90", "w")
+TEST_MODULE_NAME = "test"
+test_file.write("\nmodule " + TEST_MODULE_NAME + "\n" + TABS +
+             "use CODATA_constants\n" + TABS +
+             "use functions, only: write_CODATA\n" + TABS +
+             "use, intrinsic :: iso_fortran_env, only: wp=>real64\n" +
+             TABS + "implicit none\n\ncontains\n" + TABS + "subroutine tests\n")
+
 #-------------------------------------------------------------------------------
 # Format description of the NIST allascii.txt file:
 #
@@ -89,14 +136,15 @@ for line in lines_list:
         if line2 == 125*'-':
             header = False
     else:
-        # Extracting the four columns:
-        quantity = line2[0:59].strip()
-        value = line2[60:84].strip()
-        uncertainty = line2[85:109].strip()
+        # Extracting the four columns
+        # (with Python the right bound is excluded):
+        quantity = line2[0:60].strip()
+        value = line2[60:85].strip()
+        uncertainty = line2[85:110].strip()
         unit = line2[110:]
 
         # A special 2014 case:  "{220} lattice spacing of silicon"
-        # a variable name must begin by a letter
+        # => a variable name must begin by a letter
         quantity = quantity.replace("{220}", "twotwozero")
         # Variable names must contain only letters, underscores and numerals.
         # => all other characters will be replaced by underscores.
@@ -116,12 +164,20 @@ for line in lines_list:
         if value.find(".") == -1 and value.find("e") == -1:
             value = value + "e0"
 
-        # Adding working precision suffix:
-        value = value + "_wp"
+        # Test: print the value in a file
+        result = get_number_parts(value)
+        print(value, result)
+        left = result[0]
+        right = result[1]
+        expo = result[2]
+        test_file.write(TABS + "call write_CODATA('"+line2[0:60]+"', &\n" + 
+                        2*TABS +"& " +quantity+", "+str(left)+
+                        ", "+str(right)+", "+str(expo)+")\n")
 
         # Fortran declaration:
         f_file.write(TABS + "real(wp), parameter :: " + quantity + " = " + value
-                     + " ! " + uncertainty + "  " + unit + "\n")
+                     + "_wp ! " + uncertainty + "  " + unit + "\n")
+
         nb_constants = nb_constants + 1
 
 # Finalizing the output file:
@@ -130,3 +186,7 @@ f_file.close()
 
 print(str(nb_constants)+" constants written")
 print('{}{}'.format(os.stat("src/"+F_FILE_NAME).st_size, " bytes"))
+
+# Finalizing the test file:
+test_file.write("end subroutine tests\nend module "+TEST_MODULE_NAME+"\n")
+test_file.close()