+++ test/main.f90
@@ -1,6 +1,7 @@
 program main
     use CODATA_constants, only: c=>speed_of_light_in_vacuum, Avogadro_constant, &
         & Boltzmann_constant, electron_volt
+    use test
     implicit none
 
     print *, "Do you know this constant with so many zeros?"
@@ -11,4 +12,7 @@ program main
     print *, Avogadro_constant
     print *, Boltzmann_constant
     print *, electron_volt
+    print *, "********* all constants ************"
+    call tests
+
 end program main